Revert "Feature/lobot refactor"
