+++ helpers/Sherlock.js
