--- lobot/lobot.js
@@ -1,37 +1,0 @@
-import { HtmlHead } from "./html/head.js";
-
-const componentBundle = {
-    HtmlHead: HtmlHead,
-}
-
-/** Reads and returns the text content of the given resource **/
-async function LoadResource (name, type) {
-  const lobotResourcePath = `lobot/${type}/${name}.${type}`;
-
-  return (`css`)
-          ? `<style>${await Deno.readTextFile(lobotResourcePath)}</style>`
-          : (`js`)
-          ? `<script>${await Deno.readTextFile(lobotResourcePath)}</script>`
-          : (`txt`)
-          ? `<p>${await Deno.readTextFile(lobotResourcePath)}</p>`
-          : (`json`)
-          ? await Deno.readTextFile(lobotResourcePath)
-          : null;
-}
-
-async function LoadHtmlBundle () {
-  return componentBundle
-}
-
-async function Serve(resource) {
-  const name = resource.split(`:`)[0];
-  const type = resource.split(`:`)[1];
-  if (type === `css` || type === `js` || type === `txt` ) {
-    return await LoadResource(name, type);
-  }
-  if (type === `html`) {
-    return await LoadHtmlBundle(name);
-  }
-}
-
-export { Serve };
